pyproject.toml
[tool.poetry]
name = "line-ai-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
line-bot-sdk = "^3.11.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.1.44"
langchain-openai = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
